requirements.txt
requests
beautifulsoup4
openai
schedule
gradio
newspaper3k
lxml_html_clean